Embedded widget: Resolve deadlock
Previous implementation waited for UI threads to complete and used
Future<DisplayModel> instead of just a DisplayModel for the
active_content_model.

On closing an example with script-driven update of the display file,
this happened:

UI thread waiting for value of active_content_model so that it can
dispose that model.

Concurrently, script just set active_content_model to a new
RuntimeThread which is trying to dispose the previous content on the UI
thread.

--> UI thread waiting for runtime thread which in turn is waiting for a
submitted UI thread Runnable.

In update, nothing awaits the runtime threads to complete, and waits for
the UI thread use a timeout so they can - with error message - continue
without blocking forever.