For display background
Setting the "-fx-background:" of the root node propagates
to all child nodes in the scene graph.
In edit mode, this results in error messages because the linear-gradient
doesn't "work" for all nodes.
In the runtime, the background color style is applied to for example the
TextEntryRepresentation, overriding its jfx_node.setBackground(..)
setting.

Setting just the scroll body background to a plain color or grid image
provides basic color control.
In edit mode, the horiz_bound, vert_bound lines and grid provide
sufficient
visual indication of the display size.