Rules: Fix rule for "array[i].element"
Failed if rule was read from XML before array was populated and index i
was thus invalid.
Now array elements are (temporarily) created while the rule is read, and
then the actual elements are replaced once the array is read.
The array element used by the rule-reader is only used to parse the
value, it's OK if the actual array element is later replaced.