Embedded display: Handle named group that's not found in embedded dsp
Used to result in empty display. Now issues warning and then uses the
complete display.
Requesting a group without defining it is an error in the user's display
definition, but it happens in the real world and showing the complete
embedded display (oh, I forgot to define the sub-group) is better than
showing nothing (what's happening, why is there nothing?).