XYPlot: Example and warning regarding import of XYGraph
#243 : BOY XYGraph can be either for waveforms or scalars.

For waveforms, it behaves like the XYPlot and the import is quite good.

For scalars, it behaves like a Data Browser but without archive access
and with some trouble handling all received updates.

XYPlot cannot determine at edit time which type the PV will provide, so
it cannot automatically change into a Data Browser widget.