Widget property 'path'
When selecting several XYPlot widgets in the editor and then changing
the 'visible' setting of a y axis, this must update the
'y_axes[1].visible' setting and not the 'visible' property of the
widget.
The search for the corresponding property in other widgets must thus use
the complete path to a property, not just the name.

In the editor, the tool tip of a property now also displays the full
path to a property, which can help when writing scripts to access a
property.