Popover: Remove focus listener
Hiding the popover as it resp. its owner looses focus seemed like a good
idea, but in practice it's possible for the popover to loose focus
during normal operation. One example is a popover that includes a combo
or choice box with many entries. The list of choices is larger than the
popover. When user selects an entry, the mouse ends up outside the
popover as the list closes -> popover looses focus and closes, even
though the user simply selected an item that's conceptually within the
popover and intends to continue to work with the popover.