Fix "Cannot determine path" warning in rule dialog
In rules, the expression editor uses the same code as the property
editor, but not with the actual property of the widget, instead using a
clone of the property. So 'getPath' doesn't work for these cloned
properties, since they're not found in the widget.

-> No problem because getPath only needed when updating multiple
properties of the same widget. Skip that code for 'other' when there
aren't other widgets.

https://github.com/kasemir/org.csstudio.display.builder/issues/364