let colorbar_axis minimum be <1 even if it is log scale
Users wanted to be able to set log scale minimum to a very small
value to be able to shift the color bar. For example, 10^0 to 10^3
will look different than 10^-3 to 10^3.

Mininum must still be >0 so chose an arbitrary value of 10^-3.

Not tested, but I think this is the only change needed.